Add tentative WPTs for FileSystemAccess file/directory handles.
This change includes a new helper framed_test() which sets up a
variety of same- and cross-site subframes and helps with testing
across frames. It isn't specific to this storage API, so PLMK if
if there is a better common library to put this in.

Left TODOS against crbug.com/1322897 for remaining cases in these
tests.

Bug: 1322897

Change-Id: Ic380178077f5b99df21dac0777502cb754116faf