[mathml] Add rules for maction to UA stylesheet
This CL adds CSS rules in MathML user agent stylesheet for the legacy
<maction> element [1]. Currently, the fourth test case of WPT test
legacy-mrow-like-elements-002 is broken since the ref uses the same
content as the test file. Fix that by making the ref a green square.
Additionally, this adds default-properties-on-semantics-and-maction.html
to check the display values of semantics/maction's children in a more
direct way.

[1] https://w3c.github.io/mathml-core/#enlivening-expressions

Bug: 6606
Change-Id: I56d1b802cb3002330aecee35d939d2fef499e1f2